Settled on a conduit API. Moved a bunch of stuff into "coldsync.h", mainly structures and stuff that has to do with configuration; stuff that might be of interest to a bunch of files.
A bunch of changes to the conduit API; this is currently in flux: up
until now, I'd assumed that a conduit would be a dynamically-linked
object module, but now I'm convinced that it's better if conduits are
external programs. Naturally, this involves a lot of reorganization,
so the code is currently a mess.

Added declarations for new-API functions run_Fetch_conduits() and
run_Dump_conduits().