Lots of changes:
- got actual documentation from Handspring on what their vendor specific
setup commands do.  The biggest result of this is that we now use a
"portmapper" type mechanism to find the USB endpoint that does the Hot
Sync service, rather than just "knowing" that it's 2.

- Access the fields of the USB request using some accessor macros.  The
data is in little endian order, so it just works now.  But it would have
had problems on a big-endian CPU.

- moved printing of the "Press HotSync button" message into the device
specific I/O functions.  In the case of USB, this needs to be pressed
to cause the device to attach to the USB port.  Before, it would pring
the message after you pushed the button.

- reflect that the USB header file is installed in
/usr/include/dev/usb/usb.h now.  I don't know if you want to use this
in the autoconf process or not.

- picked up the string.h macros in PConnection_usb.c

- cleaned up the error handling in the USB code to not just exit()
at the slightest provocation.

- added an io_trace variable and IO_TRACE() macro to control debugging
in the USB code.  This could also be used in the serial code if
you wanted to add I/O specific debugging messages there.

- added a somewhat unsatisfactory way of specifing the device type on
the command line.  E.g., you can now say -t 3 along with -p /dev/ugen0
to indicate it's a USB device.  Certainly you'd prefer to say -t usb,
but I'm not sure how to do this and keep the config parser in sync.